# swagger.yaml
swagger: "2.0"
info:
  description: "This API facilitates the ability to track who gives and receives shoutouts at DEPT®."
  version: "1.0.0"
  title: "DEPT® Shoutouts"
host: "TBD"
basePath: "/api"
paths:
  /shoutouts/latest:
    get:
      summary: "Returns the 10 most recent shoutouts"
      tags:
        - shoutouts
      produces:
        - "application/json"
      responses:
        "200":
          description: "Successful Operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Shoutout"
        "401":
          description: "Unauthorized"
        "500":
          description: "Unexpected Error"
      security:
        - api_key: []
  /shoutouts/by-year:
    get:
      summary: "Returns all shoutouts for the given year (YYYY format) or past 12 months if no valid year given"
      tags:
        - shoutouts
      produces:
        - "application/json"
      parameters:
        - name: "year"
          in: "query"
          required: false
          type: "string"
      responses:
        "200":
          description: "Successful Operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Shoutout"
        "401":
          description: "Unauthorized"
        "500":
          description: "Unexpected Error"
      security:
        - api_key: []
  /profile/search:
    get:
      summary: "Searches for people given name or email and returns their user info"
      tags:
        - profile
      produces:
        - "application/json"
      parameters:
        - name: "name"
          in: "query"
          description: "partial or full name of a person"
          required: false
          type: "string"
        - name: "email"
          in: "query"
          description: "partial or full email address of a person"
          required: false
          type: "string"
      responses:
        "200":
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/Person"
        "400":
          description: 'Bad Request (Valid "name" or "email" Param Required)'
        "401":
          description: "Unauthorized"
        "500":
          description: "Unexpected Error"
      security:
        - api_key: []
  /profile/{_id}:
    get:
      summary: "Returns information about a person including shoutouts given and received"
      tags:
        - profile
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "_id of a person"
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/Profile"
        "401":
          description: "Unauthorized"
        "500":
          description: "Unexpected Error"
      security:
        - api_key: []
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Person:
    type: "object"
    properties:
      _id:
        type: "string"
      employeeId:
        type: "number"
      email:
        type: "string"
      team:
        type: "string"
      country:
        type: "string"
      name:
        type: "string"
  Shoutout:
    type: "object"
    properties:
      _id:
        type: "string"
      author:
        $ref: "#/definitions/Person"
      recipients:
        type: "array"
        items:
          $ref: "#/definitions/Person"
      text:
        type: "string"
      createDate:
        type: "string"
      month:
        type: "number"
      year:
        type: "number"
      inPastTwelveMonths:
        type: "boolean"
  Profile:
    type: "object"
    properties:
      userInfo:
        $ref: "#/definitions/Person"
      shoutoutsGiven:
        type: "array"
        items:
          $ref: "#/definitions/Shoutout"
      shoutoutsRecevied:
        type: "array"
        items:
          $ref: "#/definitions/Shoutout"